Move 'cooldown' related members to AACListener since all calls stem from there
